requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
requests==2.10.0
uber-rides==0.2.1
Werkzeug==0.11.10
